Move AtomicUniquePointer to a new file
