+++ libraries/lib-realtime-effects/RealtimeEffectState.h
@@ -16,6 +16,7 @@
 #include <optional>
 #include <unordered_map>
 #include <vector>
+#include "AtomicUniquePointer.h"
 #include "ClientData.h"
 #include "EffectInterface.h"
 #include "GlobalVariable.h"
